SpringRestDBApp/src/test/java/lesson2/dao/UserDaoTest.java
package lesson2.dao;

import lesson2.model.User;
import org.junit.Before;
import org.junit.Test;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.datasource.embedded.EmbeddedDatabase;
import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;

import java.util.List;

import static org.junit.Assert.assertEquals;


public class UserDaoTest {

    private UserDao dao;

    @Before
    public void setUp() {
        EmbeddedDatabase db = new EmbeddedDatabaseBuilder()
                .setType(EmbeddedDatabaseType.HSQL)
                .generateUniqueName(true)
                .addScript("dao/data.sql")
                .build();
        dao = new UserDao(new JdbcTemplate(db));
    }

    @Test
    public void read() throws Exception {
        List<User> list = dao.findAll();
        assertEquals(1, list.size());
        assertEquals("EmbeddedUserName", list.get(0).getName());

    }

}